select left

+++ drop_db.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
+source ./verification.sh
 
-function delete_db(){
+function drop_db(){
 	DBName=$1
-	
+	if ! validate_database_name "$DBName"; then
+        return 
+    fi
+
 	if [ -d "Databases/$DBName" ]
 	then
 		rm -r "Databases/$DBName"
@@ -11,4 +15,4 @@ function delete_db(){
 		echo "Database doesn't exist"
 	fi	
 }
-delete_db $1
+drop_db $1
